Complete Damas-Milner type inference.
This commit is the first commit where the Damas-Milner type inference
implementation passes all of the basic tests: type propagation through
assignment, typing for polymorphic functions, instantiation of
polymorphic functions. A test file with these tests is included in this
commit.

Although this means the basic system is in place, the interface still
has rough edges and there are probably some hidden bugs.