More fixes for typesys v0.2.0.9005.
This commit gets the package working with the aforementioned version of
typesys. The quantify function has been moved to typesys and a few
functions in that package have been renamed.

This commit also experimentally removes active variable tracking, since
it is not clear it is really needed when using SSA names.