Fix bug in rm_def.
This commit fixes a bug in the rm_def function. The cause of the bug is
the fact that the names attribute for an S4 class that extends a list is
stored on the object rather than its `.Data` slot, so removing a list
element from `.Data` may incorrectly update the names.

Thanks to @duncantl for pointing out the bug!